Finished Product list screen

+++ lib/models/Pricing.dart
@@ -1,3 +1,4 @@
+import 'dart:convert';
 class Pricing {
   double price;
   String storeId;
@@ -7,7 +8,7 @@ class Pricing {
   Pricing({this.price, this.storeId, this.options, this.storeName});
 
   Pricing.fromJson(Map<dynamic, dynamic> json) {
-    this.price = json["price"];
+    this.price = json["price"].toDouble();
     this.storeId = json["storeId"];
     this.options = json["options"];
     this.storeName = json["storeName"];
